rename utils folder to modules
