+++ packages/app/src/components/Dimensions/styles/Dimensions.style.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../utils/colors';
+import { colors } from '../../../modules/colors';
 
 export const styles = {
     divContainer: {
