+++ packages/app/src/components/Layout/styles/style.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../utils/colors';
+import { colors } from '../../../modules/colors';
 
 // Axis
 export const AXIS_PADDING = '8px 8px 2px 8px';
