+++ packages/app/src/reducers/ui.js
@@ -3,8 +3,8 @@ import {
     getItemIdsByDimension,
     getFilteredLayout,
     getSwapModObj,
-} from '../utils/layout';
-import { getOptionsForUi, getOptionsFromVisualization } from '../utils/options';
+} from '../modules/layout';
+import { getOptionsForUi, getOptionsFromVisualization } from '../modules/options';
 import { COLUMN } from '../components/VisualizationTypeSelector/visualizationTypes';
 
 export const SET_UI = 'SET_UI';
